refactor: reorganise files
