+++ inst/include/mo.hpp
@@ -1,9 +1,9 @@
 #ifndef MO_HPP
 #define MO_HPP
 
-#include <mo/stats/generator.hpp>
-#include <mo/stats/walker.hpp>
 #include <mo/math/misc.hpp>
 #include <mo/math/sets.hpp>
+#include <mo/stats/generator.hpp>
+#include <mo/stats/walker.hpp>
 
 #endif // MO_HPP
