+++ inst/include/mo/stats/generator_impl/fixeddblgenerator.ipp
@@ -1,5 +1,5 @@
-#ifndef MO_STATS_IMPLEMENTATION_FIXEDDBLGENERATOR_IPP
-#define MO_STATS_IMPLEMENTATION_FIXEDDBLGENERATOR_IPP
+#ifndef MO_STATS_GENERATOR_IMPL_FIXEDDBLGENERATOR_IPP
+#define MO_STATS_GENERATOR_IMPL_FIXEDDBLGENERATOR_IPP
 
 #include <mo/stats/generator.hpp>
 
@@ -20,4 +20,4 @@ inline double FixedDblGenerator::operator()(const double& value) const {
 } // stats
 } // mo
 
-#endif // MO_STATS_IMPLEMENTATION_FIXEDDBLGENERATOR_IPP
+#endif // MO_STATS_GENERATOR_IMPL_FIXEDDBLGENERATOR_IPP
