+++ inst/include/mo/stats/generator_impl/rpermutationgenerator.ipp
@@ -1,5 +1,5 @@
-#ifndef MO_STATS_GENERATOR_IMPLEMENTATION_RPERMUTATIONGENERATOR_IPP
-#define MO_STATS_GENERATOR_IMPLEMENTATION_RPERMUTATIONGENERATOR_IPP
+#ifndef MO_STATS_GENERATOR_IMPL_RPERMUTATIONGENERATOR_IPP
+#define MO_STATS_GENERATOR_IMPL_RPERMUTATIONGENERATOR_IPP
 
 #include <mo/stats/generator.hpp>
 
@@ -37,4 +37,4 @@ inline VECTOR RPermutationGenerator<VECTOR>::operator()() const {
 } // stats
 } // mo
 
-#endif // MO_STATS_GENERATOR_IMPLEMENTATION_RPERMUTATIONGENERATOR_IPP
+#endif // MO_STATS_GENERATOR_IMPL_RPERMUTATIONGENERATOR_IPP
