+++ inst/include/mo/stats/generator_impl/runifgenerator.ipp
@@ -1,5 +1,5 @@
-#ifndef MO_STATS_GENERATOR_IMPLEMENTATION_RUNIFGENERATOR01_IPP
-#define MO_STATS_GENERATOR_IMPLEMENTATION_RUNIFGENERATOR01_IPP
+#ifndef MO_STATS_GENERATOR_IMPL_RUNIFGENERATOR01_IPP
+#define MO_STATS_GENERATOR_IMPL_RUNIFGENERATOR01_IPP
 
 #include <Rcpp.h>
 #include <mo/stats/generator.hpp>
@@ -16,4 +16,4 @@ inline double RUnifGenerator01::operator()() const {
 } // stats
 } // mo
 
-#endif // MO_STATS_GENERATOR_IMPLEMENTATION_RUNIFGENERATOR01_IPP
+#endif // MO_STATS_GENERATOR_IMPL_RUNIFGENERATOR01_IPP
