+++ inst/include/mo/stats/generator_impl/runifintgenerator.ipp
@@ -1,5 +1,5 @@
-#ifndef MO_STATS_IMPLEMENTATION_RUNIFORMINTGENERATOR_IPP
-#define MO_STATS_IMPLEMENTATION_RUNIFORMINTGENERATOR_IPP
+#ifndef MO_STATS_GENERATOR_IMPL_RUNIFORMINTGENERATOR_IPP
+#define MO_STATS_GENERATOR_IMPL_RUNIFORMINTGENERATOR_IPP
 
 #include <mo/stats/generator.hpp>
 #include <Rinternals.h> // for R_xlen_t
@@ -21,4 +21,4 @@ inline R_xlen_t RUnifIntGenerator::operator()() const {
 } // stats
 } // mo
 
-#endif // MO_STATS_IMPLEMENTATION_RUNIFORMINTGENERATOR_IPP
+#endif // MO_STATS_GENERATOR_IMPL_RUNIFORMINTGENERATOR_IPP
