+++ inst/include/mo/stats/walker_impl/rsamplewalkernoreplace.ipp
@@ -1,5 +1,5 @@
-#ifndef MO_STATS_IMPLEMENTATION_RSAMPLEWALKERNOREPLACE_IPP
-#define MO_STATS_IMPLEMENTATION_RSAMPLEWALKERNOREPLACE_IPP
+#ifndef MO_STATS_WALKER_IMPL_RSAMPLEWALKERNOREPLACE_IPP
+#define MO_STATS_WALKER_IMPL_RSAMPLEWALKERNOREPLACE_IPP
 
 #include <vector>
 #include <R_ext/Utils.h>
@@ -70,4 +70,4 @@ inline R_xlen_t RSampleWalkerNoReplace::operator()() {
 } // stats
 } // mo
 
-#endif // MO_STATS_IMPLEMENTATION_RSAMPLEWALKERNOREPLACE_IPP
+#endif // MO_STATS_WALKER_IMPL_RSAMPLEWALKERNOREPLACE_IPP
