+++ inst/include/mo/stats/walker_impl/runifsamplewalkernoreplace.ipp
@@ -1,5 +1,5 @@
-#ifndef MO_STATS_IMPLEMENTATION_RUNIFSAMPLEWALKERNOREPLACE_IPP
-#define MO_STATS_IMPLEMENTATION_RUNIFSAMPLEWALKERNOREPLACE_IPP
+#ifndef MO_STATS_WALKER_IMPL_RUNIFSAMPLEWALKERNOREPLACE_IPP
+#define MO_STATS_WALKER_IMPL_RUNIFSAMPLEWALKERNOREPLACE_IPP
 
 #include <mo/stats/walker.hpp>
 #include <Rinternals.h> // for R_xlen_t
@@ -32,4 +32,4 @@ inline R_xlen_t RUnifSampleWalkerNoReplace::operator()() {
 } // stats
 } // mo
 
-#endif // MO_STATS_IMPLEMENTATION_RUNIFSAMPLEWALKERNOREPLACE_IPP
+#endif // MO_STATS_WALKER_IMPL_RUNIFSAMPLEWALKERNOREPLACE_IPP
